Rename BootModeMappingPage.jsx to just BootModeMapping.jsx
