+++ www/src/Locales/en/BootModeMapping.jsx
